se corrigio algo

+++ Tarea2/quicksort.cpp
@@ -1,8 +1,8 @@
 #include<bits/stdc++.h>
 using namespace std;
-typedef long int lInt;
+typedef unsigned long long int lInt;
 
 int quiksort(vector<lInt> &array){
-    
+
     return 0;
 }
